refactor(cycles): move the due-cycle clause into a query-only module
The clause helper landed in common.py, which holds constants and pure value
helpers and is imported by contract and schema modules. Adding SQLAlchemy and
Cycle imports there made those consumers inherit an ORM dependency.

Move it unchanged into brain/systems/cycles/queries.py and document the
inclusive flag: True compares next_run_at <= cutoff, False compares <.
